# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Household Finance API
  version: 0.1.0
  description: |
    HTTP API for the self-hosted household finance app. All authenticated routes expect
    `Authorization: Bearer <JWT>` unless noted. See `docs/API_INDEX.md` for prose topic guides.
servers:
  - url: /
    description: Relative to configured API base URL
...
paths:
...
  /resolution:
    get:
      tags: [resolution]
      summary: List resolution items (query status, type)
      security: [{ bearerAuth: [] }]
  /resolution/bulk:
    post:
      tags: [resolution]
      summary: Bulk update resolution item status
      security: [{ bearerAuth: [] }]
  /resolution/{id}:
    patch:
      tags: [resolution]
      summary: Update resolution item status
      security: [{ bearerAuth: [] }]
  /resolution/bulk-apply-category:
    post:
      tags: [resolution]
      summary: Bulk apply category to unknown-category items
      security: [{ bearerAuth: [] }]
  /resolution/summary:
    get:
      tags: [resolution]
      summary: Resolution queue summary counts
      security: [{ bearerAuth: [] }]
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT